added suppport for celery
