+++ .env
@@ -1,0 +1,7 @@
+db_host=127.0.0.1
+db_port=5432
+db_user_name=postgres
+db_password=test@123
+db_name=odk_data_automation
+db_schema=public
+db_table=prerna_saathi_final
