+++ requirements.txt
@@ -1,3 +1,4 @@
 Flask==2.0.1
 psycopg2-binary==2.9.1
 python-dotenv==0.18.0
+celery==4.4.1
